Add files via upload
Hello Analytics World,
This Video Mainly For Your Concern.

Here, I Do An Data Analytics Based On 50 StartUp Company Data.
Here, I Do Startups profit prediction using Multiple Linear Regression.

*we will be predicting the profit from the startup’s dataset with the features available to us. We’re using the 50-startups dataset for this problem statement and we will be using the concept of Multiple linear regression to predict the profit of startups companies.

Here, This machine learning model will be quite helpful in such a situation where we need to find a profit based on how much we are spending in the market and for the market. In a nutshell, this machine learning model will help to find out the profit based on the amount which we spend from the 50 startups dataset.

So, Come On Guys, Lets Go Through It.